+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-08-18
+### Added
+ - Added 3 transformers: `es6-module-nodes`, `camelcase-to-kebabcase`, `kebabcase-to-camelcase`.
+ - Added verbose logging via the `debug` setting.
+
 ## [1.1.0] - 2020-07-29
 ### Added
  - Reference providers.